# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WooCommerce Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryUpdateImage } from './categoryUpdateImage';

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  parent?: number;
  description?: string;
  image?: CategoryUpdateImage;
}